+++ hutool-core/src/main/java/com/xiaoleilu/hutool/util/IdcardUtil.java
@@ -1,0 +1,512 @@
+package com.xiaoleilu.hutool.util;
+
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
+
+import com.xiaoleilu.hutool.date.DateTime;
+import com.xiaoleilu.hutool.date.DateUtil;
+import com.xiaoleilu.hutool.lang.Validator;
+
+/**
+ * 身份证相关工具类<br>
+ * see https://www.oschina.net/code/snippet_1611_2881
+ * 
+ * @author Looly
+ * @since 3.0.4
+ */
+public class IdcardUtil {
+
+	/** 中国公民身份证号码最小长度。 */
+	private static final int CHINA_ID_MIN_LENGTH = 15;
+	/** 中国公民身份证号码最大长度。 */
+	private static final int CHINA_ID_MAX_LENGTH = 18;
+	/** 每位加权因子 */
+	private static final int power[] = { 7, 9, 10, 5, 8, 4, 2, 1, 6, 3, 7, 9, 10, 5, 8, 4, 2 };
+	/** 省市代码表 */
+	private static Map<String, String> cityCodes = new HashMap<String, String>();
+	/** 台湾身份首字母对应数字 */
+	private static Map<String, Integer> twFirstCode = new HashMap<String, Integer>();
+	/** 香港身份首字母对应数字 */
+	private static Map<String, Integer> hkFirstCode = new HashMap<String, Integer>();
+	
+	static {
+		cityCodes.put("11", "北京");
+		cityCodes.put("12", "天津");
+		cityCodes.put("13", "河北");
+		cityCodes.put("14", "山西");
+		cityCodes.put("15", "内蒙古");
+		cityCodes.put("21", "辽宁");
+		cityCodes.put("22", "吉林");
+		cityCodes.put("23", "黑龙江");
+		cityCodes.put("31", "上海");
+		cityCodes.put("32", "江苏");
+		cityCodes.put("33", "浙江");
+		cityCodes.put("34", "安徽");
+		cityCodes.put("35", "福建");
+		cityCodes.put("36", "江西");
+		cityCodes.put("37", "山东");
+		cityCodes.put("41", "河南");
+		cityCodes.put("42", "湖北");
+		cityCodes.put("43", "湖南");
+		cityCodes.put("44", "广东");
+		cityCodes.put("45", "广西");
+		cityCodes.put("46", "海南");
+		cityCodes.put("50", "重庆");
+		cityCodes.put("51", "四川");
+		cityCodes.put("52", "贵州");
+		cityCodes.put("53", "云南");
+		cityCodes.put("54", "西藏");
+		cityCodes.put("61", "陕西");
+		cityCodes.put("62", "甘肃");
+		cityCodes.put("63", "青海");
+		cityCodes.put("64", "宁夏");
+		cityCodes.put("65", "新疆");
+		cityCodes.put("71", "台湾");
+		cityCodes.put("81", "香港");
+		cityCodes.put("82", "澳门");
+		cityCodes.put("91", "国外");
+		
+		twFirstCode.put("A", 10);
+		twFirstCode.put("B", 11);
+		twFirstCode.put("C", 12);
+		twFirstCode.put("D", 13);
+		twFirstCode.put("E", 14);
+		twFirstCode.put("F", 15);
+		twFirstCode.put("G", 16);
+		twFirstCode.put("H", 17);
+		twFirstCode.put("J", 18);
+		twFirstCode.put("K", 19);
+		twFirstCode.put("L", 20);
+		twFirstCode.put("M", 21);
+		twFirstCode.put("N", 22);
+		twFirstCode.put("P", 23);
+		twFirstCode.put("Q", 24);
+		twFirstCode.put("R", 25);
+		twFirstCode.put("S", 26);
+		twFirstCode.put("T", 27);
+		twFirstCode.put("U", 28);
+		twFirstCode.put("V", 29);
+		twFirstCode.put("X", 30);
+		twFirstCode.put("Y", 31);
+		twFirstCode.put("W", 32);
+		twFirstCode.put("Z", 33);
+		twFirstCode.put("I", 34);
+		twFirstCode.put("O", 35);
+		
+		hkFirstCode.put("A", 1);
+		hkFirstCode.put("B", 2);
+		hkFirstCode.put("C", 3);
+		hkFirstCode.put("R", 18);
+		hkFirstCode.put("U", 21);
+		hkFirstCode.put("Z", 26);
+		hkFirstCode.put("X", 24);
+		hkFirstCode.put("W", 23);
+		hkFirstCode.put("O", 15);
+		hkFirstCode.put("N", 14);
+	}
+
+	/**
+	 * 将15位身份证号码转换为18位
+	 * 
+	 * @param idCard 15位身份编码
+	 * @return 18位身份编码
+	 */
+	public static String convert15To18(String idCard) {
+		StringBuilder idCard18;
+		if (idCard.length() != CHINA_ID_MIN_LENGTH) {
+			return null;
+		}
+		if (Validator.isNumber(idCard)) {
+			// 获取出生年月日
+			String birthday = idCard.substring(6, 12);
+			Date birthDate = DateUtil.parse(birthday, "yyMMdd");
+			// 获取出生年(完全表现形式,如：2010)
+			int sYear = DateUtil.year(birthDate);
+			idCard18 = StrUtil.builder().append(idCard.substring(0, 6)).append(sYear).append(idCard.substring(8));
+			// 获取校验位
+			char sVal = getCheckCode18(idCard18.toString());
+			idCard18.append(sVal);
+		} else {
+			return null;
+		}
+		return idCard18.toString();
+	}
+
+	/**
+	 * 是否有效身份证号
+	 * @param idCard 身份证号，支持18位、15位和港澳台的10位
+	 * @return 是否有效
+	 */
+	public static boolean isValidCard(String idCard) {
+		idCard = idCard.trim();
+		int length = idCard.length();
+		switch (length) {
+			case 18://18位身份证
+				return isvalidCard18(idCard);
+			case 15://15位身份证
+				return isvalidCard15(idCard);
+			case 10:{//10位身份证，港澳台地区
+				String[] cardval = isValidCard10(idCard);
+				if (cardval != null && cardval[2].equals("true")) {
+					return true;
+				}else{
+					return false;
+				}
+			}
+			default:
+				return false;
+		}
+	}
+
+	/**
+	 * 验证18位身份编码是否合法
+	 * 
+	 * @param idCard 身份编码
+	 * @return 是否合法
+	 */
+	public static boolean isvalidCard18(String idCard) {
+		if (CHINA_ID_MAX_LENGTH != idCard.length()) {
+			return false;
+		}
+		
+		// 前17位
+		String code17 = idCard.substring(0, 17);
+		// 第18位
+		char code18 = Character.toLowerCase(idCard.charAt(17));
+		if (Validator.isNumber(code17)) {
+			// 获取校验位
+			char val = getCheckCode18(code17);
+			if (val == code18) {
+				return true;
+			}
+		}
+		return false;
+	}
+
+	/**
+	 * 验证15位身份编码是否合法
+	 * 
+	 * @param idCard 身份编码
+	 * @return 是否合法
+	 */
+	public static boolean isvalidCard15(String idCard) {
+		if (CHINA_ID_MIN_LENGTH != idCard.length()) {
+			return false;
+		}
+		if (Validator.isNumber(idCard)) {
+			//省份
+			String proCode = idCard.substring(0, 2);
+			if (null == cityCodes.get(proCode)) {
+				return false;
+			}
+			
+			//生日
+			DateTime birthDate = DateUtil.parse(idCard.substring(6, 12), "yyMMdd");
+			if (false == Validator.isBirthday(birthDate.year(), birthDate.month(), birthDate.dayOfMonth())) {
+				return false;
+			}
+		} else {
+			return false;
+		}
+		return true;
+	}
+
+	/**
+	 * 验证10位身份编码是否合法
+	 * 
+	 * @param idCard 身份编码
+	 * @return 身份证信息数组
+	 *         <p>
+	 *         [0] - 台湾、澳门、香港 [1] - 性别(男M,女F,未知N) [2] - 是否合法(合法true,不合法false) 若不是身份证件号码则返回null
+	 *         </p>
+	 */
+	public static String[] isValidCard10(String idCard) {
+		String[] info = new String[3];
+		String card = idCard.replaceAll("[\\(|\\)]", "");
+		if (card.length() != 8 && card.length() != 9 && idCard.length() != 10) {
+			return null;
+		}
+		if (idCard.matches("^[a-zA-Z][0-9]{9}$")) { // 台湾
+			info[0] = "台湾";
+			System.out.println("11111");
+			String char2 = idCard.substring(1, 2);
+			if (char2.equals("1")) {
+				info[1] = "M";
+				System.out.println("MMMMMMM");
+			} else if (char2.equals("2")) {
+				info[1] = "F";
+				System.out.println("FFFFFFF");
+			} else {
+				info[1] = "N";
+				info[2] = "false";
+				System.out.println("NNNN");
+				return info;
+			}
+			info[2] = isValidTWCard(idCard) ? "true" : "false";
+		} else if (idCard.matches("^[1|5|7][0-9]{6}\\(?[0-9A-Z]\\)?$")) { // 澳门
+			info[0] = "澳门";
+			info[1] = "N";
+		} else if (idCard.matches("^[A-Z]{1,2}[0-9]{6}\\(?[0-9A]\\)?$")) { // 香港
+			info[0] = "香港";
+			info[1] = "N";
+			info[2] = isValidHKCard(idCard) ? "true" : "false";
+		} else {
+			return null;
+		}
+		return info;
+	}
+
+	/**
+	 * 验证台湾身份证号码
+	 * 
+	 * @param idCard 身份证号码
+	 * @return 验证码是否符合
+	 */
+	public static boolean isValidTWCard(String idCard) {
+		String start = idCard.substring(0, 1);
+		String mid = idCard.substring(1, 9);
+		String end = idCard.substring(9, 10);
+		Integer iStart = twFirstCode.get(start);
+		Integer sum = iStart / 10 + (iStart % 10) * 9;
+		char[] chars = mid.toCharArray();
+		Integer iflag = 8;
+		for (char c : chars) {
+			sum +=Integer.valueOf(String.valueOf(c)) * iflag;
+			iflag--;
+		}
+		return (sum % 10 == 0 ? 0 : (10 - sum % 10)) == Integer.valueOf(end) ? true : false;
+	}
+
+	/**
+	 * 验证香港身份证号码(存在Bug，部份特殊身份证无法检查)
+	 * <p>
+	 * 身份证前2位为英文字符，如果只出现一个英文字符则表示第一位是空格，对应数字58 前2位英文字符A-Z分别对应数字10-35 最后一位校验码为0-9的数字加上字符"A"，"A"代表10
+	 * </p>
+	 * <p>
+	 * 将身份证号码全部转换为数字，分别对应乘9-1相加的总和，整除11则证件号码有效
+	 * </p>
+	 * 
+	 * @param idCard 身份证号码
+	 * @return 验证码是否符合
+	 */
+	public static boolean isValidHKCard(String idCard) {
+		String card = idCard.replaceAll("[\\(|\\)]", "");
+		Integer sum = 0;
+		if (card.length() == 9) {
+			sum = (Integer.valueOf(card.substring(0, 1).toUpperCase().toCharArray()[0]) - 55) * 9 + (Integer.valueOf(card.substring(1, 2).toUpperCase().toCharArray()[0]) - 55) * 8;
+			card = card.substring(1, 9);
+		} else {
+			sum = 522 + (Integer.valueOf(card.substring(0, 1).toUpperCase().toCharArray()[0]) - 55) * 8;
+		}
+		String mid = card.substring(1, 7);
+		String end = card.substring(7, 8);
+		char[] chars = mid.toCharArray();
+		Integer iflag = 7;
+		for (char c : chars) {
+			sum = sum + Integer.valueOf(String.valueOf(c)) * iflag;
+			iflag--;
+		}
+		if (end.toUpperCase().equals("A")) {
+			sum += 10;
+		} else {
+			sum += Integer.valueOf(end);
+		}
+		return (sum % 11 == 0) ? true : false;
+	}
+
+	/**
+	 * 根据身份编号获取生日
+	 * 
+	 * @param idCard 身份编号
+	 * @return 生日(yyyyMMdd)
+	 */
+	public static String getBirthByIdCard(String idCard) {
+		Integer len = idCard.length();
+		if (len < CHINA_ID_MIN_LENGTH) {
+			return null;
+		} else if (len == CHINA_ID_MIN_LENGTH) {
+			idCard = convert15To18(idCard);
+		}
+		return idCard.substring(6, 14);
+	}
+	
+	/**
+	 * 根据身份编号获取年龄
+	 * 
+	 * @param idCard 身份编号
+	 * @return 年龄
+	 */
+	public static int getAgeByIdCard(String idCard) {
+		return getAgeByIdCard(idCard, DateUtil.date());
+	}
+	
+	/**
+	 * 根据身份编号获取指定日期当时的年龄年龄
+	 * 
+	 * @param idCard 身份编号
+	 * @param dateToCompare 以此日期为界，计算年龄。
+	 * @return 年龄
+	 */
+	public static int getAgeByIdCard(String idCard, Date dateToCompare) {
+		String birth = getBirthByIdCard(idCard);
+		return DateUtil.age(DateUtil.parse(birth, "yyyyMMdd"), dateToCompare);
+	}
+
+	/**
+	 * 根据身份编号获取生日年
+	 * 
+	 * @param idCard 身份编号
+	 * @return 生日(yyyy)
+	 */
+	public static Short getYearByIdCard(String idCard) {
+		Integer len = idCard.length();
+		if (len < CHINA_ID_MIN_LENGTH) {
+			return null;
+		} else if (len == CHINA_ID_MIN_LENGTH) {
+			idCard = convert15To18(idCard);
+		}
+		return Short.valueOf(idCard.substring(6, 10));
+	}
+
+	/**
+	 * 根据身份编号获取生日月
+	 * 
+	 * @param idCard 身份编号
+	 * @return 生日(MM)
+	 */
+	public static Short getMonthByIdCard(String idCard) {
+		Integer len = idCard.length();
+		if (len < CHINA_ID_MIN_LENGTH) {
+			return null;
+		} else if (len == CHINA_ID_MIN_LENGTH) {
+			idCard = convert15To18(idCard);
+		}
+		return Short.valueOf(idCard.substring(10, 12));
+	}
+
+	/**
+	 * 根据身份编号获取生日天
+	 * 
+	 * @param idCard 身份编号
+	 * @return 生日(dd)
+	 */
+	public static Short getDateByIdCard(String idCard) {
+		Integer len = idCard.length();
+		if (len < CHINA_ID_MIN_LENGTH) {
+			return null;
+		} else if (len == CHINA_ID_MIN_LENGTH) {
+			idCard = convert15To18(idCard);
+		}
+		return Short.valueOf(idCard.substring(12, 14));
+	}
+
+	/**
+	 * 根据身份编号获取性别
+	 * 
+	 * @param idCard 身份编号
+	 * @return 性别(1-男，0-女，-1-未知)
+	 */
+	public static int getGenderByIdCard(String idCard) {
+		int sGender = -1;
+		if (idCard.length() == CHINA_ID_MIN_LENGTH) {
+			idCard = convert15To18(idCard);
+		}
+		char sCardChar = idCard.charAt(16);
+		if (Integer.parseInt(String.valueOf(sCardChar)) % 2 != 0) {
+			sGender = 1;
+		} else {
+			sGender = 0;
+		}
+		return sGender;
+	}
+
+	/**
+	 * 根据身份编号获取户籍省份
+	 * 
+	 * @param idCard 身份编码
+	 * @return 省级编码。
+	 */
+	public static String getProvinceByIdCard(String idCard) {
+		int len = idCard.length();
+		if (len == CHINA_ID_MIN_LENGTH || len == CHINA_ID_MAX_LENGTH) {
+			String sProvinNum = idCard.substring(0, 2);
+			return cityCodes.get(sProvinNum);
+		}
+		return null;
+	}
+	
+	//----------------------------------------------------------------------------------- Private method start
+	/**
+	 * 获得18位身份证校验码
+	 * @param code17 18位身份证号中的前17位 
+	 * @return 第18位
+	 */
+	private static char getCheckCode18(String code17){
+		int sum = getPowerSum(code17.toCharArray());
+		return getCheckCode18(sum);
+	}
+	
+	/**
+	 * 将power和值与11取模获得余数进行校验码判断
+	 * 
+	 * @param iSum
+	 * @return 校验位
+	 */
+	private static char getCheckCode18(int iSum) {
+		char sCode = StrUtil.C_SPACE;
+		switch (iSum % 11) {
+			case 10:
+				sCode = '2';
+				break;
+			case 9:
+				sCode = '3';
+				break;
+			case 8:
+				sCode = '4';
+				break;
+			case 7:
+				sCode = '5';
+				break;
+			case 6:
+				sCode = '6';
+				break;
+			case 5:
+				sCode = '7';
+				break;
+			case 4:
+				sCode = '8';
+				break;
+			case 3:
+				sCode = '9';
+				break;
+			case 2:
+				sCode = 'x';
+				break;
+			case 1:
+				sCode = '0';
+				break;
+			case 0:
+				sCode = '1';
+				break;
+		}
+		return sCode;
+	}
+	
+	/**
+	 * 将身份证的每位和对应位的加权因子相乘之后，再得到和值
+	 * 
+	 * @param iArr
+	 * @return 身份证编码。
+	 */
+	private static int getPowerSum(char[] iArr) {
+		int iSum = 0;
+		if (power.length == iArr.length) {
+			for (int i = 0; i < iArr.length; i++) {
+				iSum += Integer.valueOf(String.valueOf(iArr[i])) * power[i];
+			}
+		}
+		return iSum;
+	}
+	//----------------------------------------------------------------------------------- Private method end
+}