+++ hutool-core/src/test/java/com/xiaoleilu/hutool/core/convert/ConvertTest.java
@@ -6,7 +6,6 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import com.xiaoleilu.hutool.convert.Convert;
-import com.xiaoleilu.hutool.util.ArrayUtil;
 import com.xiaoleilu.hutool.util.CharsetUtil;
 
 /**
@@ -32,11 +31,36 @@ public class ConvertTest {
 	public void toIntArrayTest() {
 		String[] b = { "1", "2", "3", "4" };
 		
-		Integer[] intArray = Convert.toIntArray(false, (Object[])b);
+		Integer[] intArray = Convert.toIntArray(b);
 		Assert.assertArrayEquals(intArray, new Integer[]{1,2,3,4});
 		
-		long[] c = {1,2,3,4,5};
-		Convert.toIntArray(false, (Object[])ArrayUtil.wrap(c));
+		long[] c = {1L,2L,3L,4L,5L};
+		Integer[] intArray2 = Convert.toIntArray(c);
+		Assert.assertArrayEquals(intArray2, new Integer[]{1,2,3,4,5});
+	}
+	
+	@Test
+	public void toLongArrayTest() {
+		String[] b = { "1", "2", "3", "4" };
+		
+		Long[] intArray = Convert.toLongArray(b);
+		Assert.assertArrayEquals(intArray, new Long[]{1L,2L,3L,4L});
+		
+		int[] c = {1,2,3,4,5};
+		Long[] intArray2 = Convert.toLongArray(c);
+		Assert.assertArrayEquals(intArray2, new Long[]{1L,2L,3L,4L,5L});
+	}
+	
+	@Test
+	public void toDoubleArrayTest() {
+		String[] b = { "1", "2", "3", "4" };
+		
+		Double[] intArray = Convert.toDoubleArray(b);
+		Assert.assertArrayEquals(intArray, new Double[]{1D,2D,3D,4D});
+		
+		int[] c = {1,2,3,4,5};
+		Double[] intArray2 = Convert.toDoubleArray(c);
+		Assert.assertArrayEquals(intArray2, new Double[]{1D,2D,3D,4D,5D});
 	}
 	
 	@Test